+++ harfbuzz/src/hb-buffer-deserialize-text.hh
@@ -30,6 +30,7 @@
 #define HB_BUFFER_DESERIALIZE_TEXT_HH
 
 #include "hb.hh"
+#include "hb-font.hh"
 
 
 #line 36 "hb-buffer-deserialize-text.hh"
@@ -378,9 +379,7 @@ _resume:
 	case 5:
 #line 55 "hb-buffer-deserialize-text.rl"
 	{
-	if (!hb_font_glyph_from_string (font,
-					tok, p - tok,
-					&info.codepoint))
+	if (!font->glyph_from_string (tok, p - tok, &info.codepoint))
 	  return false;
 }
 	break;
@@ -414,9 +413,7 @@ _resume:
 	case 4:
 #line 55 "hb-buffer-deserialize-text.rl"
 	{
-	if (!hb_font_glyph_from_string (font,
-					tok, p - tok,
-					&info.codepoint))
+	if (!font->glyph_from_string (tok, p - tok, &info.codepoint))
 	  return false;
 }
 #line 43 "hb-buffer-deserialize-text.rl"
@@ -491,9 +488,7 @@ _again:
 	case 4:
 #line 55 "hb-buffer-deserialize-text.rl"
 	{
-	if (!hb_font_glyph_from_string (font,
-					tok, p - tok,
-					&info.codepoint))
+	if (!font->glyph_from_string (tok, p - tok, &info.codepoint))
 	  return false;
 }
 #line 43 "hb-buffer-deserialize-text.rl"