+++ harfbuzz/src/hb-ot-shape-complex-hangul.cc
@@ -126,7 +126,7 @@ is_zero_width_char (hb_font_t *font,
 		    hb_codepoint_t unicode)
 {
   hb_codepoint_t glyph;
-  return hb_font_get_glyph (font, unicode, 0, &glyph) && hb_font_get_glyph_h_advance (font, glyph) == 0;
+  return hb_font_get_glyph (font, unicode, 0, &glyph) && font->get_glyph_h_advance (glyph) == 0;
 }
 
 static void