+++ CHANGELOG.md
@@ -31,3 +31,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Deprecated methods.
 - `hb_unicode_funcs_t`, since we are using only internal Unicode tables.
 - `hb_font_funcs_t`, since we are using internal OpenType parser.
+- Most of the `hb_font_t` query API.